src/requirements.txt
Flask
flask-restful
markdown
mysql-connector-python
requests
jsonpickle
timedelta
sentry-sdk[flask]